requirements.txt
pyrogram==2.0.106
pymongo==4.6.1
google-generativeai==0.3.2
python-dotenv==1.0.0
Pillow==10.2.0
gtts==2.4.0
requests==2.31.0
schedule==1.2.0
pytz==2023.4
motor==3.3.2
aiofiles==23.2.1
asyncio-mqtt==0.16.1
